+++ include/guik/camera/topdown_camera_control.hpp
@@ -15,8 +15,8 @@ public:
 
   virtual void reset_center() override;
   virtual void lookat(const Eigen::Vector3f& pt) override;
-  virtual void mouse(const Eigen::Vector2i& p, int button, bool down) override;
-  virtual void drag(const Eigen::Vector2i& p, int button) override;
+  virtual void mouse(const Eigen::Vector2f& p, int button, bool down) override;
+  virtual void drag(const Eigen::Vector2f& p, int button) override;
   virtual void scroll(const Eigen::Vector2f& rel) override;
   virtual Eigen::Vector2f depth_range() const override;
 
@@ -31,7 +31,7 @@ protected:
   Eigen::Vector3f center;
   double distance;
 
-  Eigen::Vector2i drag_last_pos;
+  Eigen::Vector2f drag_last_pos;
 
   bool left_button_down;
   bool middle_button_down;
